Criando tela de gerenciamento de regras para retenção de SKU
- Neste commit também foi feito ajuste na consulta que retorna os SKUs a serem limpos e no script que irá gerar os SKUs

+++ apps/adm-api/src/jobs/jobLimparProdutosLogistica.php
@@ -14,6 +14,10 @@ return new class extends AbstractJob {
         $produtos = LogisticaItemModel::listarProdutosLogisticasLimpar();
 
         foreach ($produtos as $produto) {
+            if (!$produto['esta_expirado']) {
+                continue;
+            }
+
             $produtoLogistica = (new ProdutoLogistica())->newFromBuilder($produto);
 
             $produtoLogistica->deleteOrFail();
